--- src/types/common.ts
@@ -1,4 +1,0 @@
-import nodeFetch from 'node-fetch';
-
-export type FetchClient = typeof nodeFetch | typeof window.fetch
-export type SearchParams = Record<string, any>
