+++ src/types/hooks.ts
@@ -1,6 +1,6 @@
 import {HTTPError} from '../errors';
 import {IsoRequest, IsoResponse} from './globals';
-import {Request, RequestConfig} from './options';
+import {Request, RequestConfig} from './request';
 
 export type BeforeRequestHook = (request: IsoRequest, config: RequestConfig) => IsoRequest | IsoResponse | void | Promise<IsoRequest | IsoResponse | void>;
 export type AfterResponseHook = (request: Request, config: RequestConfig, response: IsoResponse) => IsoResponse | void | Promise<IsoResponse | void>;
@@ -19,19 +19,4 @@ export interface Hooks {
      * запустится сразу после ответа (4хх, 5хх) на запрос
      */
     beforeError?: BeforeErrorHook[]
-}
-
-export type HookStaticMethods = {
-    beforeRequest: {
-        use: (hook: BeforeRequestHook) => BeforeRequestHook
-        remove: (hook: BeforeRequestHook) => void
-    }
-    afterResponse: {
-        use: (hook: AfterResponseHook) => AfterResponseHook
-        remove: (hook: AfterResponseHook) => void
-    }
-    beforeError: {
-        use: (hook: BeforeErrorHook) => BeforeErrorHook
-        remove: (hook: BeforeErrorHook) => void
-    }
 }
