+++ src/types/index.ts
@@ -1,6 +1,6 @@
 export * from './hooks';
 export * from './options';
 export * from './services';
-export * from './common';
 export * from './globals';
-export {HTTPClient} from './instance';
+export * from './request';
+export {HTTPClientInstance} from './instance';
