+++ src/types/options.ts
@@ -1,8 +1,4 @@
-import {
-    IsoFetch,
-    IsoHeaders,
-    IsoRequestInit,
-} from './globals';
+import {IsoFetch, IsoHeaders} from './globals';
 import {Hooks} from './hooks';
 import {CacheService} from './services';
 
@@ -23,8 +19,7 @@ export type TimerType = {
     value: any
 }
 
-type Timeout = number | null
-
+export type Timeout = number | null
 export type SearchParamsInit = string | string[][] | Record<string, string> | URLSearchParams | undefined;
 export type SearchParamsOption = SearchParamsInit | Record<string, string | number | boolean> | Array<Array<string | number | boolean>>;
 
@@ -43,39 +38,4 @@ export type NormalizedOptions = CreateInstanceOptions & {
     hooks: Required<Hooks>
 }
 
-export type RequestOptions = IsoRequestInit & {
-    prefix?: string
-    query?: SearchParamsOption
-    timeout?: Timeout
-    credentials?: 'include' | 'omit' | 'same-origin'
-    agent?: any
-    trace?: boolean
-    cache?: CacheOptions
-    hooks?: Hooks
-}
-
 export type InstanceOptions = Omit<NormalizedOptions, 'hooks' | 'cacheService'>
-
-export type RequestConfig = {
-    baseUrl: string
-    prefix: string
-    query?: SearchParamsOption
-    cache?: CacheOptions
-    trace: boolean
-}
-
-export type Request = {
-    url: string
-    headers: IsoHeaders
-    timeout: Timeout
-    method: string
-    credentials: string
-    body: any
-    signal: AbortSignal
-}
-
-export type RequestContext = {
-    request: Request
-    config: RequestConfig
-    hooks: Required<Hooks>
-}
